requirements.txt
# AWS SDK
boto3>=1.26.0

# MQTT Client
paho-mqtt>=1.6.1

# Environment Variables
python-dotenv>=0.19.0

# Testing
pytest>=7.0.0

# Code Quality
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0

# Additional dependencies
requests>=2.26.0 